+++ flash/image/classification/__init__.py
@@ -1,2 +1,2 @@
-from flash.image.classification.data import ImageClassificationData, ImageClassificationPreprocess  # noqa: F401
+from flash.image.classification.data import ImageClassificationData, ImageClassificationInputTransform  # noqa: F401
 from flash.image.classification.model import ImageClassifier  # noqa: F401