+++ flash/image/segmentation/data.py
@@ -33,7 +33,8 @@ from flash.core.data.data_source import (
     PathsDataSource,
     TensorDataSource,
 )
-from flash.core.data.process import Deserializer, Preprocess
+from flash.core.data.io.input_transform import InputTransform
+from flash.core.data.process import Deserializer
 from flash.core.data.utils import image_default_loader
 from flash.core.utilities.imports import (
     _FIFTYONE_AVAILABLE,
@@ -217,7 +218,7 @@ class SemanticSegmentationDeserializer(ImageDeserializer):
         return result
 
 
-class SemanticSegmentationPreprocess(Preprocess):
+class SemanticSegmentationInputTransform(InputTransform):
     def __init__(
         self,
         train_transform: Optional[Dict[str, Callable]] = None,
@@ -230,7 +231,7 @@ class SemanticSegmentationPreprocess(Preprocess):
         labels_map: Dict[int, Tuple[int, int, int]] = None,
         **data_source_kwargs: Any,
     ) -> None:
-        """Preprocess pipeline for semantic segmentation tasks.
+        """InputTransform pipeline for semantic segmentation tasks.
 
         Args:
             train_transform: Dictionary with the set of transforms to apply during training.
@@ -291,7 +292,7 @@ class SemanticSegmentationPreprocess(Preprocess):
 class SemanticSegmentationData(DataModule):
     """Data module for semantic segmentation tasks."""
 
-    preprocess_cls = SemanticSegmentationPreprocess
+    input_transform_cls = SemanticSegmentationInputTransform
 
     @staticmethod
     def configure_data_fetcher(
@@ -316,19 +317,19 @@ class SemanticSegmentationData(DataModule):
         test_transform: Optional[Dict[str, Callable]] = None,
         predict_transform: Optional[Dict[str, Callable]] = None,
         data_fetcher: Optional[BaseDataFetcher] = None,
-        preprocess: Optional[Preprocess] = None,
+        input_transform: Optional[InputTransform] = None,
         val_split: Optional[float] = None,
         batch_size: int = 4,
         num_workers: int = 0,
-        **preprocess_kwargs: Any,
+        **input_transform_kwargs: Any,
     ) -> "DataModule":
 
-        if "num_classes" not in preprocess_kwargs:
+        if "num_classes" not in input_transform_kwargs:
             raise MisconfigurationException("`num_classes` should be provided during instantiation.")
 
-        num_classes = preprocess_kwargs["num_classes"]
+        num_classes = input_transform_kwargs["num_classes"]
 
-        labels_map = getattr(preprocess_kwargs, "labels_map", None) or SegmentationLabels.create_random_labels_map(
+        labels_map = getattr(input_transform_kwargs, "labels_map", None) or SegmentationLabels.create_random_labels_map(
             num_classes
         )
 
@@ -348,11 +349,11 @@ class SemanticSegmentationData(DataModule):
             test_transform=test_transform,
             predict_transform=predict_transform,
             data_fetcher=data_fetcher,
-            preprocess=preprocess,
+            input_transform=input_transform,
             val_split=val_split,
             batch_size=batch_size,
             num_workers=num_workers,
-            **preprocess_kwargs,
+            **input_transform_kwargs,
         )
 
         if dm.train_dataset is not None:
@@ -374,13 +375,13 @@ class SemanticSegmentationData(DataModule):
         test_transform: Optional[Dict[str, Callable]] = None,
         predict_transform: Optional[Dict[str, Callable]] = None,
         data_fetcher: Optional[BaseDataFetcher] = None,
-        preprocess: Optional[Preprocess] = None,
+        input_transform: Optional[InputTransform] = None,
         val_split: Optional[float] = None,
         batch_size: int = 4,
         num_workers: int = 0,
         num_classes: Optional[int] = None,
         labels_map: Dict[int, Tuple[int, int, int]] = None,
-        **preprocess_kwargs,
+        **input_transform_kwargs,
     ) -> "DataModule":
         """Creates a :class:`~flash.image.segmentation.data.SemanticSegmentationData` object from the given data
         folders and corresponding target folders.
@@ -397,25 +398,25 @@ class SemanticSegmentationData(DataModule):
                 corresponding inputs).
             predict_folder: The folder containing the predict data.
             train_transform: The dictionary of transforms to use during training which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             val_transform: The dictionary of transforms to use during validation which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             test_transform: The dictionary of transforms to use during testing which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             predict_transform: The dictionary of transforms to use during predicting which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             data_fetcher: The :class:`~flash.core.data.callback.BaseDataFetcher` to pass to the
                 :class:`~flash.core.data.data_module.DataModule`.
-            preprocess: The :class:`~flash.core.data.data.Preprocess` to pass to the
-                :class:`~flash.core.data.data_module.DataModule`. If ``None``, ``cls.preprocess_cls``
+            input_transform: The :class:`~flash.core.data.data.InputTransform` to pass to the
+                :class:`~flash.core.data.data_module.DataModule`. If ``None``, ``cls.input_transform_cls``
                 will be constructed and used.
             val_split: The ``val_split`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
             batch_size: The ``batch_size`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
             num_workers: The ``num_workers`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
             num_classes: Number of classes within the segmentation mask.
             labels_map: Mapping between a class_id and its corresponding color.
-            preprocess_kwargs: Additional keyword arguments to use when constructing the preprocess. Will only be used
-                if ``preprocess = None``.
+            input_transform_kwargs: Additional keyword arguments to use when constructing the input_transform.
+                Will only be used if ``input_transform = None``.
 
         Returns:
             The constructed data module.
@@ -438,13 +439,13 @@ class SemanticSegmentationData(DataModule):
             test_transform=test_transform,
             predict_transform=predict_transform,
             data_fetcher=data_fetcher,
-            preprocess=preprocess,
+            input_transform=input_transform,
             val_split=val_split,
             batch_size=batch_size,
             num_workers=num_workers,
             num_classes=num_classes,
             labels_map=labels_map,
-            **preprocess_kwargs,
+            **input_transform_kwargs,
         )