+++ flash/pointcloud/detection/data.py
@@ -5,7 +5,8 @@ from torch.utils.data import Sampler
 from flash.core.data.base_viz import BaseDataFetcher
 from flash.core.data.data_module import DataModule
 from flash.core.data.data_source import BaseDataFormat, DataSource, DefaultDataKeys, DefaultDataSources
-from flash.core.data.process import Deserializer, Preprocess
+from flash.core.data.io.input_transform import InputTransform
+from flash.core.data.process import Deserializer
 from flash.pointcloud.detection.open3d_ml.data_sources import (
     PointCloudObjectDetectionDataFormat,
     PointCloudObjectDetectorFoldersDataSource,
@@ -35,7 +36,7 @@ class PointCloudObjectDetectorDatasetDataSource(DataSource):
         }
 
 
-class PointCloudObjectDetectorPreprocess(Preprocess):
+class PointCloudObjectDetectorInputTransform(InputTransform):
     def __init__(
         self,
         train_transform: Optional[Dict[str, Callable]] = None,
@@ -72,7 +73,7 @@ class PointCloudObjectDetectorPreprocess(Preprocess):
 
 class PointCloudObjectDetectorData(DataModule):
 
-    preprocess_cls = PointCloudObjectDetectorPreprocess
+    input_transform_cls = PointCloudObjectDetectorInputTransform
 
     @classmethod
     def from_folders(
@@ -86,7 +87,7 @@ class PointCloudObjectDetectorData(DataModule):
         test_transform: Optional[Dict[str, Callable]] = None,
         predict_transform: Optional[Dict[str, Callable]] = None,
         data_fetcher: Optional[BaseDataFetcher] = None,
-        preprocess: Optional[Preprocess] = None,
+        input_transform: Optional[InputTransform] = None,
         val_split: Optional[float] = None,
         batch_size: int = 4,
         num_workers: int = 0,
@@ -95,12 +96,12 @@ class PointCloudObjectDetectorData(DataModule):
         labels_folder_name: Optional[str] = "labels",
         calibrations_folder_name: Optional[str] = "calibs",
         data_format: Optional[BaseDataFormat] = PointCloudObjectDetectionDataFormat.KITTI,
-        **preprocess_kwargs: Any,
+        **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given folders using the
         :class:`~flash.core.data.data_source.DataSource` of name
         :attr:`~flash.core.data.data_source.DefaultDataSources.FOLDERS`
-        from the passed or constructed :class:`~flash.core.data.process.Preprocess`.
+        from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
             train_folder: The folder containing the train data.
@@ -108,24 +109,24 @@ class PointCloudObjectDetectorData(DataModule):
             test_folder: The folder containing the test data.
             predict_folder: The folder containing the predict data.
             train_transform: The dictionary of transforms to use during training which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             val_transform: The dictionary of transforms to use during validation which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             test_transform: The dictionary of transforms to use during testing which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             predict_transform: The dictionary of transforms to use during predicting which maps
-                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+                :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             data_fetcher: The :class:`~flash.core.data.callback.BaseDataFetcher` to pass to the
                 :class:`~flash.core.data.data_module.DataModule`.
-            preprocess: The :class:`~flash.core.data.data.Preprocess` to pass to the
-                :class:`~flash.core.data.data_module.DataModule`. If ``None``, ``cls.preprocess_cls``
+            input_transform: The :class:`~flash.core.data.data.InputTransform` to pass to the
+                :class:`~flash.core.data.data_module.DataModule`. If ``None``, ``cls.input_transform_cls``
                 will be constructed and used.
             val_split: The ``val_split`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
             batch_size: The ``batch_size`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
             num_workers: The ``num_workers`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
             sampler: The ``sampler`` to use for the ``train_dataloader``.
-            preprocess_kwargs: Additional keyword arguments to use when constructing the preprocess. Will only be used
-                if ``preprocess = None``.
+            input_transform_kwargs: Additional keyword arguments to use when constructing the input_transform.
+                Will only be used if ``input_transform = None``.
             scans_folder_name: The name of the pointcloud scan folder
             labels_folder_name: The name of the pointcloud scan labels folder
             calibrations_folder_name: The name of the pointcloud scan calibration folder
@@ -154,7 +155,7 @@ class PointCloudObjectDetectorData(DataModule):
             test_transform=test_transform,
             predict_transform=predict_transform,
             data_fetcher=data_fetcher,
-            preprocess=preprocess,
+            input_transform=input_transform,
             val_split=val_split,
             batch_size=batch_size,
             num_workers=num_workers,
@@ -163,5 +164,5 @@ class PointCloudObjectDetectorData(DataModule):
             labels_folder_name=labels_folder_name,
             calibrations_folder_name=calibrations_folder_name,
             data_format=data_format,
-            **preprocess_kwargs,
+            **input_transform_kwargs,
         )